backend_python/requirements_selfhosted.txt
# GameSpot Booking System - SELF-HOSTED AI STACK
# NO QUOTAS, NO LIMITS, COMPLETELY FREE

# ═══════════════════════════════════════════════════════════════
# CORE WEB FRAMEWORK
# ═══════════════════════════════════════════════════════════════
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# ═══════════════════════════════════════════════════════════════
# DATABASE
# ═══════════════════════════════════════════════════════════════
mysql-connector-python==8.2.0

# ═══════════════════════════════════════════════════════════════
# SECURITY & ENVIRONMENT
# ═══════════════════════════════════════════════════════════════
bcrypt==4.1.2
python-dotenv==1.0.0
python-dateutil==2.8.2

# ═══════════════════════════════════════════════════════════════
# 🚀 SELF-HOSTED AI STACK (UNLIMITED, FREE)
# ═══════════════════════════════════════════════════════════════

# LLM Engine - Mistral-7B-Instruct (Local, Unlimited)
torch==2.1.2
transformers==4.36.2
accelerate==0.25.0
sentencepiece==0.1.99

# Speech-to-Text - Whisper (OpenAI, Open-source)
openai-whisper==20231117

# Text-to-Speech - Coqui TTS (Local, Open-source)
TTS==0.22.0

# Audio Processing
numpy==1.24.3
soundfile==0.12.1

# ═══════════════════════════════════════════════════════════════
# FALLBACK VOICE (Always available)
# ═══════════════════════════════════════════════════════════════
gTTS==2.5.0

# ═══════════════════════════════════════════════════════════════
# OPTIONAL: Lightweight Models (Faster, less RAM)
# ═══════════════════════════════════════════════════════════════
# If 7B models are too slow, use these instead:
# - microsoft/phi-2 (2.7B params)
# - TinyLlama/TinyLlama-1.1B-Chat-v1.0 (1.1B params)

# ═══════════════════════════════════════════════════════════════
# REMOVED DEPENDENCIES (No longer needed)
# ═══════════════════════════════════════════════════════════════
# google-generativeai - REMOVED (had quotas)
# edge-tts - REMOVED (unreliable)
# openai - REMOVED (paid API)
# elevenlabs - REMOVED (paid API)

# ═══════════════════════════════════════════════════════════════
# SYSTEM REQUIREMENTS
# ═══════════════════════════════════════════════════════════════
# RAM: 8GB minimum (16GB recommended)
# Storage: 15GB for models (one-time download)
# GPU: Optional (CUDA speeds up inference 10x)
# OS: Linux, macOS, Windows (all supported)

# ═══════════════════════════════════════════════════════════════
# INSTALLATION NOTES
# ═══════════════════════════════════════════════════════════════
# 
# First time setup (downloads models):
#   pip install -r requirements_selfhosted.txt
#   python -m services.selfhosted_llm_service  # Downloads Mistral-7B
#   python -m services.selfhosted_voice_service  # Downloads Whisper + Coqui
#
# Models are cached locally (~15GB total):
#   - Mistral-7B: ~/.cache/huggingface/hub/ (~14GB)
#   - Whisper base: ~/.cache/whisper/ (~140MB)
#   - Coqui TTS: ~/.local/share/tts/ (~500MB)
#
# After first run, NO internet needed!
